Code Optimization, cleanup, refactoring and commenting.
Fixes:
Method HtmlArray::fixSpaces() replaced leading space characters with the
wrong amount of '&#xA0;' characters instead of the amount of spaces as
defined in HtmlArray::options.
This method is removed and replacement of these spaces has moved to
method HtmlArray::formatLines() wich replaces the leading spaces with
the correct amount of non-breaking spaces (character '&nbsp;').